// src/api/SimuladorAdapter/CustosAdapter/CustosChart.ts
export interface CustosChart {
    Id: number;
    Data: string;
    Ponto: string;
    Posto: string;
    TipoDemanda: string;
    Demanda: number;
    TipoContrato: string;
    Contrato: number;
    Tarifa: number;
    Confiabilidade: number;
    Piu: number;
    Add: number;
    Pis: number;
    Eust: number;
    Total: number
}